-dodano w CustomerController metodę public void initializeDatabase(), która jest wywoływana po wpisaniu w url: /fillbase i dodaje samochody do tabeli -zmieniono port na 8089 (w application.properties)

+++ src/main/java/com/mechanicproject/entity/Repair.java
@@ -26,7 +26,7 @@ public class Repair {
     private LocalDate dateofEnd;
 
     @NotNull
-    private float quotation;
+    private double quotation;
 
     public Repair(Integer customerId, Integer carId, LocalDate dateOfBegin, LocalDate dateofEnd, float quotation) {
         this.customerId = customerId;
@@ -76,7 +76,7 @@ public class Repair {
         this.dateofEnd = dateofEnd;
     }
 
-    public float getQuotation() {
+    public double getQuotation() {
         return quotation;
     }
